requirements.txt
streamlit>=1.24.0
openai
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
nltk>=3.8.1
langchain>=0.0.267
PyPDF2>=3.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv
